+++ renku-core-client/src/main/scala/io/renku/core/client/RenkuCoreClient.scala
@@ -22,52 +22,38 @@ import cats.effect.Async
 import cats.syntax.all._
 import com.typesafe.config.{Config, ConfigFactory}
 import io.renku.control.Throttler
-import io.renku.graph.model.projects
-import io.renku.http.client.{AccessToken, RestClient}
-import org.http4s.Header
+import io.renku.graph.model.versions.SchemaVersion
+import io.renku.http.client.RestClient
 import org.http4s.client.dsl.Http4sClientDsl
 import org.http4s.dsl.Http4sDsl
 import org.typelevel.log4cats.Logger
-import org.typelevel.ci._
 
 trait RenkuCoreClient[F[_]] {
-  def getMigrationCheck(projectGitHttpUrl: projects.GitHttpUrl,
-                        accessToken:       AccessToken
-  ): F[Result[ProjectMigrationCheck]]
+  def findCoreUri(schemaVersion: SchemaVersion): F[Result[RenkuCoreUri.Versioned]]
 }
 
 object RenkuCoreClient {
   def apply[F[_]: Async: Logger](config: Config = ConfigFactory.load): F[RenkuCoreClient[F]] =
-    for {
-      coreCurrentUri <- RenkuCoreUri.Current.loadFromConfig[F](config)
-      versionClient = RenkuCoreVersionClient[F](coreCurrentUri, config)
-    } yield new RenkuCoreClientImpl[F](coreCurrentUri, versionClient, ClientTools[F])
+    RenkuCoreUri.Current.loadFromConfig[F](config).map { coreCurrentUri =>
+      new RenkuCoreClientImpl[F](coreCurrentUri, RenkuCoreUri.ForSchema, LowLevelApis[F](coreCurrentUri), config)
+    }
 }
 
 private class RenkuCoreClientImpl[F[_]: Async: Logger](currentUri: RenkuCoreUri.Current,
-                                                       versionClient: RenkuCoreVersionClient[F],
-                                                       clientTools:   ClientTools[F]
+                                                       coreUriForSchemaLoader: RenkuCoreUri.ForSchemaLoader,
+                                                       lowLevelApis:           LowLevelApis[F],
+                                                       config:                 Config
 ) extends RestClient[F, Nothing](Throttler.noThrottling)
     with RenkuCoreClient[F]
     with Http4sDsl[F]
     with Http4sClientDsl[F] {
 
-  import clientTools._
+  println(currentUri)
 
-  println(versionClient)
+  override def findCoreUri(schemaVersion: SchemaVersion): F[Result[RenkuCoreUri.Versioned]] =
+    for {
+      uriForSchema   <- coreUriForSchemaLoader.loadFromConfig[F](schemaVersion, config)
+      apiVersionsRes <- lowLevelApis.getApiVersion(uriForSchema)
+    } yield apiVersionsRes.map(_.max).map(RenkuCoreUri.Versioned(uriForSchema, _))
 
-  override def getMigrationCheck(projectGitHttpUrl: projects.GitHttpUrl,
-                                 accessToken:       AccessToken
-  ): F[Result[ProjectMigrationCheck]] = {
-
-    val uri = (currentUri.uri / "renku" / "cache.migrations_check")
-      .withQueryParam("git_url", projectGitHttpUrl.value)
-
-    send(GET(uri).withHeaders(Header.Raw(ci"gitlab-token", accessToken.value))) {
-      case (Ok, _, resp) =>
-        toResult[ProjectMigrationCheck](resp)
-      case reqInfo =>
-        toFailure[ProjectMigrationCheck](s"Migration check for $projectGitHttpUrl failed")(reqInfo)
-    }
-  }
 }
